made everything just needs testing
