+++ react-quiz/src/components/Nav.js
@@ -8,6 +8,8 @@ class Nav extends React.Component {
           <li onClick={() => {this.props.handleView('home')}}>Home</li>
           <li onClick={() => {this.props.handleView('viewQuizzes')}}>Browse Quizzes</li>
           <li onClick={() => {this.props.handleView('makeQuiz')}}>Create Quiz</li>
+          <li onClick={() => {this.props.handleView('updateQuiz')}}>Update Quiz</li>
+          <li onClick={() => {this.props.handleView('deleteQuiz')}}>Delete Quiz</li>
         </ul>
       </nav>
     )
