--- src/components/Header.js
@@ -1,13 +1,0 @@
-import React from 'react'
-
-class Header extends React.Component {
-  render() {
-    return(
-      <header>
-        <h1>Insert header shit here</h1>
-      </header>
-    )
-  }
-}
-
-export default Header;
